misc: add support for IB700 module
Patchwork-id: 36803
O-Subject: [RHEL6 PATCH] misc: add support for IB700 module
Bugzilla: 667090
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=667090

KVM guests emulate this simple kernel watchdog to handle scenarios
when the guest crashes.  It was discovered that RHEL-6 doesn't
compile this, therefore the guest falls back to a software watchdog
mechanism.  The guests really need something in the kernel.

Nothing has changed in this driver in a couple of years.  The
maintaince is really low and it makes things easier for the KVM
guys.

Enabled per request on 32-bit and 64-bit.

Cheers,
Don

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>