hpsa: adjust timing of post-reset sleeps
Patchwork-id: 36808
O-Subject: [RHEL6.2 PATCH 02/27] hpsa: adjust timing of post-reset sleeps.
Bugzilla: 698268
RH-Acked-by: Rob Evers <revers@redhat.com>

Wait longer for no-ops to complete (this prevents messages about retries).
Also, don't sleep as long after the reset is triggered if the doorbell
reset is used.  On newer controllers sleeping too long can cause us to miss
the transition to NOT READY, and think that the board has failed to reset.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>