cciss: do a better job of detecting controller reset failure
Patchwork-id: 36819
O-Subject: [RHEL6.2 PATCH 12/27] cciss: do a better job of detecting controller
	reset failure
Bugzilla: 698268
RH-Acked-by: Rob Evers <revers@redhat.com>

Detect failure of controller reset by noticing if the 32 bytes of
"driver version" we store on the hardware in the config table
fail to get zeroed out.  Previously we noticed if the controller
did not transition to "simple mode", but this did not detect reset
failure if the controller was already in simple mode prior to
the reset attempt (e.g. due to module parameter hpsa_simple_mode=1).

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>