cciss: use cmd_alloc not cmd_special_alloc for the kdump soft reset command
Patchwork-id: 36829
O-Subject: [RHEL6.2 PATCH 25/27] cciss: use cmd_alloc not cmd_special_alloc for
	the kdump soft reset command
Bugzilla: 698268
RH-Acked-by: Rob Evers <revers@redhat.com>

When doing the soft reset for kdump, use cmd_alloc, not cmd_special_alloc
and don't free the command.  The ensures the DMA of the command down to the
controller can complete with the command mapped for DMA all the while.  It
doesn't leak, because shortly after the reset is done, we re-initialize
everything anyway.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>