Fix resync hang after surprise removal
Patchwork-id: 36837
O-Subject: [RHEL6 PATCH] Fix resync hang after surprise removal
Bugzilla: 707268
RH-Acked-by: Rik van Riel <riel@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=707268

This patch fixes a bug we stumbled upon: after performing a surprise removal
of one member of a non-boot 2-disk raid1 pair, md would go into an infinite
resync loop.  The driver thought there was a spare available when in fact
there was none.  Stratus posted a proposed patch to linux-raid, but Neil Brown
responded with this one-liner that did the trick.  Presumably this is the
patch that will go upstream.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>