for -Werror, check compiler version instead of platform
Patchwork-id: 36838
O-Subject: [RHEL6 PATCH] [misc] for -Werror,
	check compiler version instead of platform
Bugzilla:

There was a check put into the Makefile a while ago that said if you
are not building on a RHEL6 then compiler warnings are ok.

The intent of this change was to allow developers to use later Fedora
distros and compile RHEL-6 using a later gcc which was more strict
(that led to more warnings).

However, I noticed that when using my locally compile RHEL-6 gcc on a
F15 box, the warnings failed to stop the build and I was accidentally
missing some compiler warnings when I wasn't paying attention.  This
could lead to a grumpy kernel maintainer if the warnings went unchecked!

Considering the original problem was to limit the damage done by future
compilers and that RHEL-6 will not move past gcc-4.4 for safety reasons,
I modified the Makefile check to look at the compiler version instead
of the platform.

This allows my F15 box using a RHEL-6 gcc compiler to properly fail.

Cheers,
Don