GFS2: Incorrect inode state during deallocation
Patchwork-id: 36840
O-Subject: [RHEL 6.2] GFS2: Incorrect inode state during deallocation (bz #712139)
Bugzilla: 712139
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This small patch fragment somehow got missed out of a backport of
code in both RHEL5 and the upstream kernel. As a result, when the
responsibility for deallocation is passed from one node to another
the receiving node may not have a fully uptodate inode state. If the
sending node has changed the important parts of the state in the
mean time (block allocation/deallocation) then this results in
triggering an assert during the deallocation on the receiving
node. This is easily fixed with fsck and the only data lost will
be the file that was being deallocated in the first place.

This has been tested by customers and QE and all have reported that it
resolves the problem. It is also upstream and in RHEL5.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>