cifs: wrap received signature check in srv_mutex
Patchwork-id: 36870
O-Subject: [RHEL6.2 PATCH 28/34] BZ#662626: cifs: wrap received signature check in srv_mutex
Bugzilla: 662626
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

While testing my patchset to fix asynchronous writes, I hit a bunch
of signature problems when testing with signing on. The problem seems
to be that signature checks on receive can be running at the same
time as a process that is sending, or even that multiple receives can
be checking signatures at the same time, clobbering the same data
structures.

While we're at it, clean up the comments over cifs_calculate_signature
and add a note that the srv_mutex should be held when calling this
function.

This patch seems to fix the problems for me, but I'm not clear on
whether it's the best approach. If it is, then this should probably
go to stable too.

Cc: stable@kernel.org
Cc: Shirish Pargaonkar <shirishpargaonkar@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 157c249114508aa71daa308a426e15d81a4eed00

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>