cifs: correctly handle NULL tcon pointer in CIFSTCon
Patchwork-id: 36864
O-Subject: [RHEL6.2 PATCH 34/34] BZ#662626: cifs: correctly handle NULL tcon pointer in CIFSTCon
Bugzilla: 662626
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Long ago (in commit 00e485b0), I added some code to handle share-level
passwords in CIFSTCon. That code ignored the fact that it's legit to
pass in a NULL tcon pointer when connecting to the IPC$ share on the
server.

This wasn't really a problem until recently as we only called CIFSTCon
this way when the server returned -EREMOTE. With the introduction of
commit c1508ca2 however, it gets called this way on every mount, causing
an oops when share-level security is in effect.

Fix this by simply treating a NULL tcon pointer as if user-level
security were in effect. I'm not aware of any servers that protect the
IPC$ share with a specific password anyway. Also, add a comment to the
top of CIFSTCon to ensure that we don't make the same mistake again.

Cc: <stable@kernel.org>
Reported-by: Martijn Uffing <mp3project@sarijopen.student.utwente.nl>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 8d1bca328b7c17af33bcf966d799c556ecbf370f

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>