NFSv4: Ensure we request the ordinary fileid when doing readdirplus
Patchwork-id: 36889
O-Subject: [RHEL6.2 PATCH 01/12] NFSv4: Ensure we request the ordinary fileid when doing readdirplus
Bugzilla: 714740
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

When readdir() returns a directory entry for the root of a mounted
filesystem, Linux follows the old convention of returning the inode
number of the covered directory (despite newer versions of POSIX declaring
that this is a bug).
To ensure this continues to work, the NFSv4 readdir implementation requests
the 'mounted-on-fileid' from the server.

However, readdirplus also needs to instantiate an inode for this entry, and
for that, we also need to request the real fileid as per this patch.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>

Upstream commit 28331a46d88459788c8fca72dbb0415cd7f514c9

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>