cfq-iosched: Set group_isolation=1 by default
Patchwork-id: 36907
O-Subject: [PATCH RHEL6.2 V2] cfq-iosched: Set group_isolation=1 by default
	(Was: Re: [PATCH RHEL6.2] cfq-iosched: Set group_isolation=1 by
	default)
Bugzilla: 707757
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

BZ: 707757

This is not upstream as upstream as got rid of this tunable altogether.
I am not removing this tunable in rhel6 due to ABI concerns.

Current default is 0 and it creates two problems.

- Practically all the users set group_isolation=1. without this, in
  most of the workloads we don't see service differentiation between
  groups.

- For guest IO, it can put some threads in root group and some in
  guests's cgroup. Splitting threads leads to one thread not
  preempting other thread and leads to greater idling hence reducing
  throughput.

So upstream has made this behavior default and got rid of this tunable.
I am retaining the tunable but changing the default to 1.

v2: fix documentation

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>