debug: reintroduce would_have_oomkilled procfs ctl
Patchwork-id: 36987
O-Subject: [RHEL6.2 PATCH BZ645770 V2 64/65] Fedora debug: reintroduce
	would_have_oomkilled procfs ctl
Bugzilla: 645770
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>

Bugzilla: 645770
https://bugzilla.redhat.com/show_bug.cgi?id=645770

Description:
Reintroduce of the would_have_oomkilled, which was removed(reverted) at the
beginning of this patchset. It was temporary reverted so it does not
collide with upstream oom killer patches which were backported.

revert description:
-----------------------------------8<-----------------------------------------
This feature is not in upstream and since __oom_kill_task is removed(renamed and
completely changed), I temporary reverted this commit so it does not collide
with upstream patches which are backported. I hope this will allow better review
since a lot of upstream patches modifying oomk will be a clean git-apply and we
will avoid to take this feature through all the upstream patches touching the
oom_kill_task. This feature is reintroduced after the oom killer is in sync with
upstream.
-----------------------------------8<-----------------------------------------

original description:
-----------------------------------8<-----------------------------------------
    [Fedora] [debug] add would_have_oomkilled procfs ctl

    Message-id:
    Patchwork-id:
    O-Subject:
    Bugzilla:

    Original Fedora patch: linux-2.6-debug-vm-would-have-oomkilled.patch

    This patch adds "would_have_oomkilled" control to procfs, which if enabeld
    will just print the process that would have been killed instead of actually
    killing it.

    Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>
-----------------------------------8<-----------------------------------------

Brew build: 3421026
https://brewweb.devel.redhat.com/taskinfo?taskID=3421026

Test status:
Basic functionality tested by me.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>