cifs: no need to mark smb_ses_list as cifs_demultiplex_thread is exiting
Patchwork-id: 37022
O-Subject: [RHEL6.2 PATCH 02/30] BZ#654198: cifs: no need to mark smb_ses_list as cifs_demultiplex_thread is exiting
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The TCP_Server_Info is refcounted and every SMB session holds a
reference to it. Thus, smb_ses_list is always going to be empty when
cifsd is coming down. This is dead code.

Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Reviewed-by: Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 9d78315b03fc91228306db42edc533efa69cb518
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>