cifs: clean up accesses to midCount
Patchwork-id: 37016
O-Subject: [RHEL6.2 PATCH 04/30] BZ#654198: cifs: clean up accesses to midCount
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

It's an atomic_t and the code accesses the "counter" field in it directly
instead of using atomic_read(). It also is sometimes accessed under a
spinlock and sometimes not. Move it out of the spinlock since we don't need
belt-and-suspenders for something that's just informational.

Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Reviewed-by: Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 8097531a5cb55c6472118da094dc88caf9be66ac
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>