cifs: move locked sections out of DeleteMidQEntry and AllocMidQEntry
Patchwork-id: 37011
O-Subject: [RHEL6.2 PATCH 05/30] BZ#654198: cifs: move locked sections out of DeleteMidQEntry and AllocMidQEntry
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

In later patches, we're going to need to have finer-grained control
over the addition and removal of these structs from the pending_mid_q
and we'll need to be able to call the destructor while holding the
spinlock. Move the locked sections out of both routines and into
the callers. Fix up current callers of DeleteMidQEntry to call a new
routine that dequeues the entry and then destroys it.

Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Reviewed-by: Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: ddc8cf8fc718587a3788330bf4f32b379f08b250
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>