cifs: wait indefinitely for responses
Patchwork-id: 37015
O-Subject: [RHEL6.2 PATCH 07/30] BZ#654198: cifs: wait indefinitely for responses
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The client should not be timing out on individual SMB requests. Too much
of the state between client and server is tied to the state of the
socket. If we time out requests and issue spurious disconnects then that
comprimises data integrity.

Instead of doing this complicated dance where we try to decide how long
to wait for a response for particular requests, have the client instead
wait indefinitely for a response. Also, use a TASK_KILLABLE sleep here
so that fatal signals will break out of this waiting.

Later patches will add support for detecting dead peers and forcing
reconnects based on that.

Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Reviewed-by:  Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 0ade640e9cda805692dbf688f4bb69e94719275a
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>