cifs: don't reconnect server when we don't get a response
Patchwork-id: 37012
O-Subject: [RHEL6.2 PATCH 08/30] BZ#654198: cifs: don't reconnect server when we don't get a response
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

We only want to force a reconnect to the server under very limited and
specific circumstances. Now that we have processes waiting indefinitely
for responses, we shouldn't reach this point unless a reconnect is
already in process. Thus, there's no reason to re-mark the server for
reconnect here.

Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Reviewed-by:  Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: dad255b182363db1d1124458cd3fb0a4deac0d0f
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>