cifs: reconnect unresponsive servers
Patchwork-id: 37023
O-Subject: [RHEL6.2 PATCH 14/30] BZ#654198: cifs: reconnect unresponsive servers
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Steve French <sfrench@us.ibm.com>

If the server isn't responding to echoes, we don't want to leave tasks
hung waiting for it to reply. At that point, we'll want to reconnect
so that soft mounts can return an error to userspace quickly.

If the client hasn't received a reply after a specified number of echo
intervals, assume that the transport is down and attempt to reconnect
the socket.

The number of echo_intervals to wait before attempting to reconnect is
tunable via a module parameter. Setting it to 0, means that the client
will never attempt to reconnect. The default is 5.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Upstream-commit: fda3594362184383e73f0a2a5fa5b38ac0e04fd8
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>