cifs: send an NT_CANCEL request when a process is signalled
Patchwork-id: 37029
O-Subject: [RHEL6.2 PATCH 18/30] BZ#654198: cifs: send an NT_CANCEL request when a process is signalled
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Use the new send_nt_cancel function to send an NT_CANCEL when the
process is delivered a fatal signal. This is a "best effort" enterprise
however, so don't bother to check the return code. There's nothing we
can reasonably do if it fails anyway.

Reviewed-by: Pavel Shilovsky <piastryyy@gmail.com>
Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 2db7c5815555d8daabf7d4ab1253ce690852c140
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>