cifs: force a reconnect if there are too many MIDs in flight
Patchwork-id: 37032
O-Subject: [RHEL6.2 PATCH 20/30] BZ#654198: cifs: force a reconnect if there are too many MIDs in flight
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Currently, we allow the pending_mid_q to grow without bound with
SIGKILL'ed processes. This could eventually be a DoS'able problem. An
unprivileged user could a process that does a long-running call and then
SIGKILL it.

If he can also intercept the NT_CANCEL calls or the replies from the
server, then the pending_mid_q could grow very large, possibly even to
2^16 entries which might leave GetNextMid in an infinite loop. Fix this
by imposing a hard limit of 32k calls per server. If we cross that
limit, set the tcpStatus to CifsNeedReconnect to force cifsd to
eventually reconnect the socket and clean out the pending_mid_q.

While we're at it, clean up the function a bit and eliminate an
unnecessary NULL pointer check.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Shirish Pargaonkar <shirishpargaonkar@gmail.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 92a4e0f0169498867ecb19c2244510dd4beba149
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>