cifs: fix length checks in checkSMB
Patchwork-id: 37031
O-Subject: [RHEL6.2 PATCH 21/30] BZ#654198: cifs: fix length checks in checkSMB
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The cERROR message in checkSMB when the calculated length doesn't match
the RFC1001 length is incorrect in many cases. It always says that the
RFC1001 length is bigger than the SMB, even when it's actually the
reverse.

Fix the error message to say the reverse of what it does now when the
SMB length goes beyond the end of the received data. Also, clarify the
error message when the RFC length is too big. Finally, clarify the
comments to show that the 512 byte limit on extra data at the end of
the packet is arbitrary.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 6284644e8de1f4005166c918c3d2aa4c510ab9f6
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>