cifs: don't send an echo request unless NegProt has been done
Patchwork-id: 37034
O-Subject: [RHEL6.2 PATCH 24/30] BZ#654198: cifs: don't send an echo request unless NegProt has been done
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

When the socket to the server is disconnected, the client more or less
immediately calls cifs_reconnect to reconnect the socket. The NegProt
and SessSetup however are not done until an actual call needs to be
made.

With the addition of the SMB echo code, it's possible that the server
will initiate a disconnect on an idle socket. The client will then
reconnect the socket but no NegotiateProtocol request is done. The
SMBEcho workqueue job will then eventually pop, and an SMBEcho will be
sent on the socket. The server will then reject it since no NegProt was
done.

The ideal fix would be to either have the socket not be reconnected
until we plan to use it, or to immediately do a NegProt when the
reconnect occurs. The code is not structured for this however. For now
we must just settle for not sending any echoes until the NegProt is
done.

Reported-by: JG <jg@cms.ac>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 247ec9b418ba50c9022280035330059364d54540
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>