cifs: don't always drop malformed replies on the floor (try #3)
Patchwork-id: 37038
O-Subject: [RHEL6.2 PATCH 29/30] BZ#654198: cifs: don't always drop malformed replies on the floor (try #3)
Bugzilla: 654198
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Slight revision to this patch...use min_t() instead of conditional
assignment. Also, remove the FIXME comment and replace it with the
explanation that Steve gave earlier.

After receiving a packet, we currently check the header. If it's no
good, then we toss it out and continue the loop, leaving the caller
waiting on that response.

In cases where the packet has length inconsistencies, but the MID is
valid, this leads to unneeded delays. That's especially problematic now
that the client waits indefinitely for responses.

Instead, don't immediately discard the packet if checkSMB fails. Try to
find a matching mid_q_entry, mark it as having a malformed response and
issue the callback.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: 71823baff1978be892e7a36eddf6170e1cc6650d
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>