initialise bd_super in bdget()
Patchwork-id: 37164
O-Subject: [RHEL6.2 PATCH] block: initialise bd_super in bdget()
Bugzilla: 718107
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This patch fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=718107

bd_super is currently reset to NULL in kill_block_super() so we rely on previous
users of the block_device object to initialise this value for the next user.
This quirk was exposed on RHEL5 when a third party filesystem did not always use
kill_block_super() and therefore bd_super wasn't being reset when a block_device
object was recycled within the cache.  This can cause a panic in blkdev_releasepage()
when the uninitialised bd_super field is dereferenced.

Patch has been posted upstream and the fix has been verified by customer.

Signed-off-by: Lachlan McIlroy <lmcilroy@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>