crypto: prng error in initial seed calculation.
Patchwork-id: 37172
O-Subject: [RHEL6.2 PATCH 3/3] [s390x] crypto: prng error in initial seed calculation.
Bugzilla: 709710

Description
-----------
Since the seed contains at least 8 bytes from the kernel random
pool the initial state of the random number generator is
unpredictable without knowing the seed so there is no visible
symptom.

If the prng is seeded with more than 8 bytes the offset
calculation is wrong resulting in an access beyond the
end of the kernel stack. Although the accessed data is
presumably random this weakens the quality of the inital
seed for the random number generator.

Fix the offset calculation for a seed longer than 8 bytes.

Bugzilla
--------
BZ 709710
https://bugzilla.redhat.com/show_bug.cgi?id=709710

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.39
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=ed961581a7ca91d6a4852af2e44333e983100505
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=c708c57e247775928b9a6bce7b4d8d14883bf39b

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>