elevator: fix O_DIRECT I/O starving readers
Patchwork-id: 37174
O-Subject: [rhel6 patch] elevator: fix O_DIRECT I/O starving readers
Bugzilla: 587402
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Hi,

The basic problem is that a process can issue a never-ending stream of
async direct I/Os to the same sector on a device, thus starving out
other I/O in the system (due to the way the alias handling works in
anticipatory, cfq and deadline, which service aliases before adding the
new request to the rb tree).  This patch allows aliases to be added to
the rb tree (it adds them to the right to preserve ordering, though that
doesn't matter as the order isn't guaranteed anyway).  I think this is
the preferred solution, as it doesn't break up time slices in CFQ or
batches in deadline.  I've tested it, and it does solve the starvation
issue.

Jens queued this patch into his for-3.1 branch.  This addresses bug
587402.  Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>