cifs: move "ntlmssp" and "local_leases" options out of experimental code
Patchwork-id: 37215
O-Subject: [RHEL6.2 PATCH 06/12] BZ#668791: cifs: move "ntlmssp" and "local_leases" options out of experimental code
Bugzilla: 668791
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

I see no real need to leave these sorts of options under an
EXPERIMENTAL ifdef. Since you need a mount option to turn this code
on, that only blows out the testing matrix.

local_leases has been under the EXPERIMENTAL tag for some time, but
it's only the mount option that's under this label. Move it out
from under this tag.

The NTLMSSP code is also under EXPERIMENTAL, but it needs a mount
option to turn it on, and in the future any distro will reasonably
want this enabled. Go ahead and move it out from under the
EXPERIMENTAL tag.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Acked-by: Suresh Jayaraman <sjayaraman@suse.de>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream-commit: b4d6fcf13f417464c13c6fde46e87c495ba6b6ee
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>