cifs: change bleft in decode_unicode_ssetup back to signed type
Patchwork-id: 37231
O-Subject: [RHEL6.2 PATCH 15/25] BZ#692709: cifs: change bleft in decode_unicode_ssetup back to signed type
Bugzilla: 692709
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

The buffer length checks in this function depend on this value being a
signed data type, but 690c522fa converted it to an unsigned type.

Also, eliminate a problem with the null termination check in the same
function. cifs_strndup_from_ucs handles that situation correctly
already, and the existing check could potentially lead to a buffer
overrun since it increments bleft without checking to see whether it
falls off the end of the buffer.

Cc: stable@kernel.org
Reported-and-Acked-by: David Howells <dhowells@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>