cifs: check for bytes_remaining going to zero in CIFS_SessSetup
Patchwork-id: 37240
O-Subject: [RHEL6.2 PATCH 16/25] BZ#692709: cifs: check for bytes_remaining going to zero in CIFS_SessSetup
Bugzilla: 692709
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

It's possible that when we go to decode the string area in the
SESSION_SETUP response, that bytes_remaining will be 0. Decrementing it at
that point will mean that it can go "negative" and wrap. Check for a
bytes_remaining value of 0, and don't try to decode the string area if
that's the case.

Cc: stable@kernel.org
Reported-and-Acked-by: David Howells <dhowells@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>