NLM: Don't hang forever on NLM unlock requests
Patchwork-id: 37251
O-Subject: [RHEL6.2 PATCH] BZ#709549: NLM: Don't hang forever on NLM unlock
	requests
Bugzilla: 709549
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

This bug was opened as a CVE. The basic problem is that an unprivileged
user could use this bug to trigger a rpc_task leak and DoS the box. The
following patch fixed the issue upstream, and I was able to verify it
by testing with my own reproducer as well.

Original patch description follows:

---------------------------[snip]----------------------------

If the NLM daemon is killed on the NFS server, we can currently end up
hanging forever on an 'unlock' request, instead of aborting. Basically,
if the rpcbind request fails, or the server keeps returning garbage, we
really want to quit instead of retrying.

Tested-by: Vasily Averin <vvs@sw.ru>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@kernel.org
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>