create-patches.sh: handle continuations in mail subjects
Patchwork-id: 37289
O-Subject: [RHEL6 PATCH] [redhat] create-patches.sh: handle continuations in
	mail subjects
Bugzilla:
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Newer versions of git than RHEL6 will generate continuations in the
headers (wrapping the line and indenting with a single space.) This
behaviour results in different patch names generated for the spec file
between F-15 and RHEL-6 git versions. Fix this, by handling hacking
around to append the continued line at the end of the subject. (There's
probably a better fix, but my awk-fu is weak.)

(M.O. is to flag the line we're processing, and if we get a
 newline+space before handling the next one (which will always be
 either a different $TYPE or a new patch) append it to the subject.)

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>