scsi_dh_alua: Attach to UNAVAILABLE/OFFLINE AAS devices
Patchwork-id: 37330
O-Subject: [RHEL6.2 PATCH] scsi_dh_alua: Attach to UNAVAILABLE/OFFLINE AAS
	devices
Bugzilla: 698329
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: Martin George <marting@netapp.com>

This is for bz 698329.

I have tested with the netapp sim by setting the device's
group to offline then adding disks to verify that the
module now attaches and that IO is failed when the state
is offline/unavailable, but allows IO when put in a online state.

>From the upstream commit:

The SCSI ALUA handler currently fails to attach to devices
reporting an UNAVAILABLE/OFFLINE AAS. But given that an
UNAVAILABLE/OFFLINE AAS can transition to other states
like ACTIVE/OPTIMIZED, ACTIVE/NON-OPTIMIZED, etc. as per
SPC4, this ALUA handler behavior should be rectified so
as to attach to devices which also report an
UNAVAILABLE/OFFLINE AAS.

Signed-off-by: Martin George <marting@netapp.com>
Reviewed-by: Mike Christie <michaelc@cs.wisc.edu>
Signed-off-by: James Bottomley <James.Bottomley@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>