hold the page lock until after set_page_stable_node
Patchwork-id: 37425
O-Subject: Re: [RHEL6.2 PATCH] hold the page lock until after
	set_page_stable_node
Bugzilla: 683658
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Subject: hold the page lock until after set_page_stable_node

From: Andrea Arcangeli <aarcange@redhat.com>

The original patch that fixed bug 556572 was found to have a bug: it released
the page lock too soon (allowing do_wp_page reuse the wrprotected page before
PageKsm would be set in the page->mapping). This was noticed by Marcelo by code
review of Avi's crash session on Gleb's reproducer. As usual crash and gdb are
extremely helpful and Avi found the pte/spte/page values and all were found
consistent and readonly on a Ksm page.

That could lead one guest to have pagetable and shadow pagetable write access
to a shared Ksm page. It's extremely hard to reproduce so it went unnoticed so
long.

The fix for bug 556572 was merged in RHEL6 on May 19 2010. Upstream took a long
time to merge the fix and by the time it was merged the fix was changed not to
alter the locking but the original fix wasn't claimed as unsafe so there was no
apparent reason until now to want to use the upstraem fix for bug 556572. The
new version of the fix for bug 556572 was merged upstream Jan 13 2011, more
than 6 months later RHEL6 had it fixed. And the upstream bug didn't introduce
the locking race condition.

The below patch simply reverse the original fix for bug 556572, and applies the
upstream one and it should fix bugzilla 683658 and 717197. It's good idea to
re-verify bug 556572 however to be sure the orhpaned swapcache is collected as
expected after applying this fix, to be sure not to introduce regressions.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>