ext4: Fix max file size and logical block counting of extent format file
Patchwork-id: 37530
O-Subject: [RHEL6.2 PATCH] ext4: Fix max file size and logical block counting of extent format file
Bugzilla: 722569
CVE: CVE-2011-2695
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 722569
https://bugzilla.redhat.com/show_bug.cgi?id=722569

BREW 3486860
https://brewweb.devel.redhat.com/taskinfo?taskID=3486860

This is backport of the following upstream commit. However ext4 in RHEL6
does not have punch-hole support so some RHEL specific changes were needed
to merge it.

It was tested with xfstests and stress.sh.

>From f17722f917b2f21497deb6edc62fb1683daa08e6 Mon Sep 17 00:00:00 2001
From: Lukas Czerner <lczerner@redhat.com>
Date: Mon, 6 Jun 2011 00:05:17 -0400
Subject: [PATCH 2/3] ext4: Fix max file size and logical block counting of
 extent format file

Kazuya Mio reported that he was able to hit BUG_ON(next == lblock)
in ext4_ext_put_gap_in_cache() while creating a sparse file in extent
format and fill the tail of file up to its end. We will hit the BUG_ON
when we write the last block (2^32-1) into the sparse file.

The root cause of the problem lies in the fact that we specifically set
s_maxbytes so that block at s_maxbytes fit into on-disk extent format,
which is 32 bit long. However, we are not storing start and end block
number, but rather start block number and length in blocks. It means
that in order to cover extent from 0 to EXT_MAX_BLOCK we need
EXT_MAX_BLOCK+1 to fit into len (because we counting block 0 as well) -
and it does not.

The only way to fix it without changing the meaning of the struct
ext4_extent members is, as Kazuya Mio suggested, to lower s_maxbytes
by one fs block so we can cover the whole extent we can get by the
on-disk extent format.

Also in many places EXT_MAX_BLOCK is used as length instead of maximum
logical block number as the name suggests, it is all a bit messy. So
this commit renames it to EXT_MAX_BLOCKS and change its usage in some
places to actually be maximum number of blocks in the extent.

The bug which this commit fixes can be reproduced as follows:

 dd if=/dev/zero of=/mnt/mp1/file bs=<blocksize> count=1 seek=$((2**32-2))
 sync
 dd if=/dev/zero of=/mnt/mp1/file bs=<blocksize> count=1 seek=$((2**32-1))

Reported-by: Kazuya Mio <k-mio@sx.jp.nec.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>